Moved provider in scope, added integration tests
