+++ frontend/http/src/test/java/org/dotwebstack/framework/frontend/http/provider/graph/TriGGraphProviderTest.java
@@ -1,7 +1,7 @@
-package org.dotwebstack.framework.frontend.provider.graph;
+package org.dotwebstack.framework.frontend.http.provider.graph;
 
 
-import org.dotwebstack.framework.provider.graph.TurtleGraphProvider;
+import org.dotwebstack.framework.frontend.http.provider.graph.TriGGraphProvider;
 import org.dotwebstack.framework.test.DBEERPEDIA;
 import org.eclipse.rdf4j.model.Model;
 import org.eclipse.rdf4j.model.impl.LinkedHashModel;
@@ -18,15 +18,17 @@ import org.mockito.junit.MockitoJUnitRunner;
 import javax.ws.rs.core.MediaType;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.util.List;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.ArgumentMatchers.anyInt;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @RunWith(MockitoJUnitRunner.class)
-public class TurtleGraphProviderTest {
+public class TriGGraphProviderTest {
 
     @Mock
     private OutputStream outputStream;
@@ -35,13 +37,13 @@ public class TurtleGraphProviderTest {
     private ArgumentCaptor<byte[]> byteCaptor;
 
     @Test
-    public void isWritableForTurtleMediaType() {
+    public void isWritableForTriGMediaType() {
         // Arrange
-        TurtleGraphProvider provider = new TurtleGraphProvider();
+        TriGGraphProvider provider = new TriGGraphProvider();
 
         // Act
         boolean result = provider.isWriteable(LinkedHashModel.class, null, null,
-                new MediaType("text", "turtle"));
+                new MediaType("application", "trig"));
 
         // Assert
         assertThat(result, is(true));
@@ -50,11 +52,11 @@ public class TurtleGraphProviderTest {
     @Test
     public void isNotWritableForStringClass() {
         // Arrange
-        TurtleGraphProvider provider = new TurtleGraphProvider();
+        TriGGraphProvider provider = new TriGGraphProvider();
 
         // Act
         boolean result = provider.isWriteable(String.class, null, null,
-                new MediaType("text", "turtle"));
+                new MediaType("application", "trig"));
 
         // Assert
         assertThat(result, is(false));
@@ -63,7 +65,7 @@ public class TurtleGraphProviderTest {
     @Test
     public void isNotWritableForXmlMediaType() {
         // Arrange
-        TurtleGraphProvider provider = new TurtleGraphProvider();
+        TriGGraphProvider provider = new TriGGraphProvider();
 
         // Act
         boolean result = provider.isWriteable(String.class, null, null,
@@ -74,9 +76,9 @@ public class TurtleGraphProviderTest {
     }
 
     @Test
-    public void writesTurtleFormat() throws IOException {
+    public void writesTriGFormat() throws IOException {
         // Arrange
-        TurtleGraphProvider provider = new TurtleGraphProvider();
+        TriGGraphProvider provider = new TriGGraphProvider();
         Model model = new ModelBuilder().subject(DBEERPEDIA.BREWERIES)
                 .add(RDF.TYPE, DBEERPEDIA.BACKEND)
                 .add(RDFS.LABEL, DBEERPEDIA.BREWERIES_LABEL)
@@ -86,10 +88,17 @@ public class TurtleGraphProviderTest {
         provider.writeTo(model, null, null, null, null, null, outputStream);
 
         // Assert
-        verify(outputStream).write(byteCaptor.capture(), anyInt(), anyInt());
-        String result = new String(byteCaptor.getValue());
-        assertThat(result, containsString("<http://dbeerpedia.org#Breweries> a <http://dbeerpedia.org#Backend> ;"));
-        assertThat(result, containsString("<http://www.w3.org/2000/01/rdf-schema#label> \"Beer breweries in The Netherlands\""));
+        // 2 times? feels like weird behaviour of the TriG parser
+        verify(outputStream, times(2)).write(byteCaptor.capture(), anyInt(), anyInt());
+        List<byte[]> values = byteCaptor.getAllValues();
+        String result1 = new String(values.get(0));
+        String result2 = new String(values.get(1));
+
+        assertThat(result1, containsString("<http://dbeerpedia.org#Breweries> a <http://dbeerpedia.org#Backend> ;"));
+        assertThat(result1, containsString("<http://www.w3.org/2000/01/rdf-schema#label> \"Beer breweries in The Netherlands\""));
+
+        assertThat(result2, containsString("<http://dbeerpedia.org#Breweries> a <http://dbeerpedia.org#Backend> ;"));
+        assertThat(result2, containsString("<http://www.w3.org/2000/01/rdf-schema#label> \"Beer breweries in The Netherlands\""));
     }
 
 }
