+++ integration-test/src/test/resources/openapi/dbeerpedia.yml
@@ -5,7 +5,10 @@ info:
 host: localhost
 basePath: /dbp/api/v1
 produces:
-  - text/plain
+  - text/turtle
+  - application/trig
+  - application/ld+json
+  - application/rdf+xml
 paths:
   /breweries:
     get:
